requirements.txt
python-bioformats==1.5.2
pyclowder